"Add user selection and creation to post creation form; pass users to post creation view"
This diff adds a user selection dropdown and a "create new user" button to the post creation form. It also includes a script to update the email input when a user is selected from the dropdown. Additionally, it modifies the PostController to pass a collection of all users to the post creation view.